+++ locals.tf
@@ -5,8 +5,7 @@ locals {
   enable_s3_origin                   = var.s3_origin != null ? true : false
   enable_lambda_function_association = var.lambda_function_association != null ? true : false
   resource_name                      = "${var.prefix}-${var.environment}-${var.name}-cf"
-  aliases                            = concat([var.domain_alias], var.domain_aliases_extra)
-  aliases_records                    = { for name in local.aliases : name => { "name" = name } }
+  aliases_records                    = { for name in var.domain_aliases : name => { "name" = name } }
 
   tags = merge(
     {